requirements.txt
Flask==2.3.3
requests==2.26.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.0.2
websocket-client==1.2.1
gunicorn==20.1.0
Werkzeug==2.3.7 